fixed a few gcc-side warnings and a spelling error in the name of FormRecurrentLoop() that got fixed but partially undone during the last merge; added clear installation instructions for MS MPI SDK in the source file at the place where it fails if you don't have it

+++ MachineLearning/CNTK/MPIWrapper.h
@@ -1,6 +1,8 @@
 #pragma once
 
 // This uses mpi.h which requires the Microsoft MPI SDK to be installed on Windows
+// [cf. https://msdn.microsoft.com/en-us/library/bb524831(v=vs.85).aspx]
+// download msmpisdk.msi at https://www.microsoft.com/en-us/download/details.aspx?id=47259 and run it
 // and the MPI dev package on Linux (sudo apt-get install libopenmpi-dev openmpi-bin openmpi-doc)
 #include "mpi.h"
 #pragma comment(lib, "msmpi.lib")
@@ -181,7 +183,7 @@ namespace Microsoft { namespace MSR { namespace CNTK {
             m_numNodesInUse = requestednodes;
             fprintf(stderr, "requestnodes [%s]: using %d out of %d MPI nodes (%d requested); we (%d) are %s\n",
                      msg, m_numNodesInUse, m_numMPINodes, (int) requestednodes,
-                     CurrentNodeRank(), IsIdle() ? "out (idle)" : "in (participating)");
+                     (int) CurrentNodeRank(), IsIdle() ? "out (idle)" : "in (participating)");
             fflush(stderr);
             Ping("requestnodes (after change)");
         }
